+++ protected/humhub/modules/comment/widgets/views/commentBlockedUser.php
@@ -5,22 +5,27 @@
  * @license https://www.humhub.com/licences
  */
 
+use humhub\components\View;
 use humhub\modules\comment\models\Comment;
-use humhub\modules\ui\view\components\View;
 use humhub\modules\user\widgets\Image as UserImage;
-use humhub\widgets\Button;
+use humhub\widgets\bootstrap\Button;
 
 /* @var $this View */
 /* @var $comment Comment */
 /* @var $loadBlockedCommentUrl string */
 ?>
 
-<div class="media comment-blocked-user" id="comment_<?= $comment->id; ?>"
+<div class="d-flex comment-blocked-user" id="comment_<?= $comment->id ?>"
      data-action-component="comment.Comment">
 
     <hr class="comment-separator">
 
-    <?= UserImage::widget(['user' => $comment->user, 'width' => 25, 'htmlOptions' => ['class' => 'pull-left', 'data-contentcontainer-id' => $comment->user->contentcontainer_id]]); ?>
-    <?= Yii::t('CommentModule.base', 'Comment of blocked user.') ?>
-    <?= Button::asLink(Yii::t('CommentModule.base', 'Show'))->action('showBlocked', $loadBlockedCommentUrl)->xs()->cssClass('text-primary') ?>
+    <div class="flex-shrink-0 me-2">
+        <?= UserImage::widget(['user' => $comment->user, 'width' => 25, 'htmlOptions' => ['data-contentcontainer-id' => $comment->user->contentcontainer_id]]); ?>
+    </div>
+
+    <div class="flex-grow-1">
+        <?= Yii::t('CommentModule.base', 'Comment of blocked user.') ?>
+        <?= Button::asLink(Yii::t('CommentModule.base', 'Show'))->action('showBlocked', $loadBlockedCommentUrl)->sm()->cssClass('text-primary') ?>
+    </div>
 </div>