+++ protected/humhub/modules/dashboard/views/dashboard/index_guest.php
@@ -1,6 +1,6 @@
 <?php
 
-use humhub\libs\Html;
+use humhub\helpers\Html;
 use humhub\modules\dashboard\widgets\DashboardContent;
 use humhub\modules\dashboard\widgets\Sidebar;
 use humhub\modules\space\widgets\NewSpaces;
@@ -10,10 +10,10 @@ use humhub\widgets\FooterMenu;
 
 <?= Html::beginContainer(); ?>
 <div class="row">
-    <div class="col-md-8 layout-content-container">
+    <div class="col-lg-8 layout-content-container">
         <?= DashboardContent::widget(); ?>
     </div>
-    <div class="col-md-4 layout-sidebar-container">
+    <div class="col-lg-4 layout-sidebar-container">
         <?= Sidebar::widget([
             'widgets' => [
                 [