+++ protected/humhub/modules/like/notifications/NewLike.php
@@ -9,10 +9,10 @@
 namespace humhub\modules\like\notifications;
 
 use humhub\components\ActiveRecord;
+use humhub\helpers\Html;
 use humhub\modules\content\interfaces\ContentOwner;
-use Yii;
-use yii\bootstrap\Html;
 use humhub\modules\notification\components\BaseNotification;
+use Yii;
 
 /**
  * Notifies a user about likes of his objects (posts, comments, tasks & co)