+++ protected/humhub/widgets/LinkPager.php
@@ -9,33 +9,9 @@
 namespace humhub\widgets;
 
 /**
- * @inheritdoc
+ * @deprecated since 1.18
+ * Use humhub\widgets\bootstrap\LinkPager instead
  */
-class LinkPager extends \yii\widgets\LinkPager
+class LinkPager extends \humhub\widgets\bootstrap\LinkPager
 {
-    /**
-     * @inheritdoc
-     */
-    public $maxButtonCount = 5;
-
-    /**
-     * @inheritdoc
-     */
-    public $nextPageLabel = '<i class="fa fa-step-forward"></i>';
-
-    /**
-     * @inheritdoc
-     */
-    public $prevPageLabel = '<i class="fa fa-step-backward"></i>';
-
-    /**
-     * @inheritdoc
-     */
-    public $firstPageLabel = '<i class="fa fa-fast-backward"></i>';
-
-    /**
-     * @inheritdoc
-     */
-    public $lastPageLabel = '<i class="fa fa-fast-forward"></i>';
-
 }